pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.8.1"
description = "Educational project of finding differences between two files"
authors = ["Vorobyev <nikos1997@yandex.ru>"]
readme = "README.md"
classifiers = ["Topic :: Education :: Testing",
               "Topic :: Software Development :: Libraries :: Python Modules"]
packages = [{include = "gendiff"}]
repository = 'https://github.com/proggressor/python-project-50'

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.entry_point:main"